Extract ToSonarLintIssue to CFamilyIssueToAnalysisIssueConverter
