+++ src/Integration.Vsix.UnitTests/CFamily/CFamilyHelperTests.cs
@@ -315,146 +315,6 @@ namespace SonarLint.VisualStudio.Integration.Vsix.CFamily.UnitTests
             }
         }
 
-        [TestMethod]
-        public void ToSonarLintIssue_NoMessageParts_IssueWithoutSecondaryLocations()
-        {
-            var ruleConfig = GetDummyRulesConfiguration();
-
-            var message = new Message("rule2", "file", 4, 3, 2, 1, "test endline is not zero", false, new MessagePart[0]);
-
-            // Act
-            var issue = CFamilyHelper.ToSonarLintIssue(message, "lang1", ruleConfig);
-
-            // Assert
-            issue.Flows.Should().BeEmpty();
-        }
-
-        [TestMethod]
-        public void ToSonarLintIssue_HasMessageParts_IssueWithSecondaryLocations()
-        {
-            var ruleConfig = GetDummyRulesConfiguration();
-
-            var messageParts = new List<MessagePart>
-            {
-                new MessagePart("test1.cpp", 1, 2, 3, 4, "this is a test 1"),
-                new MessagePart("test2.cpp", 5, 6, 7, 8, "this is a test 2")
-            };
-
-            var expectedLocations = new List<AnalysisIssueLocation>
-            {
-                new AnalysisIssueLocation("this is a test 2", "test2.cpp", 5, 7, 6, 8),
-                new AnalysisIssueLocation("this is a test 1", "test1.cpp", 1, 3, 2, 4)
-            };
-
-            var expectedFlows = new List<AnalysisIssueFlow>
-            {
-                new AnalysisIssueFlow(expectedLocations)
-            };
-
-            var message = new Message("rule2", "file", 4, 3, 2, 1, "test endline is not zero", false, messageParts.ToArray());
-
-            // Act
-            var issue = CFamilyHelper.ToSonarLintIssue(message, "lang1", ruleConfig);
-
-            // Assert
-            issue.Flows.Should().BeEquivalentTo(expectedFlows, x=> x.WithStrictOrdering());
-        }
-
-        [TestMethod]
-        public void ToSonarLintIssue_EndLineIsNotZero()
-        {
-            var ruleConfig = GetDummyRulesConfiguration();
-            var message = new Message("rule2", "file", 4, 3, 2, 1, "test endline is not zero", false, new MessagePart[0]);
-
-            // Act
-            var issue = CFamilyHelper.ToSonarLintIssue(message, "lang1", ruleConfig);
-
-            // Assert
-            issue.StartLine.Should().Be(4);
-            issue.StartLineOffset.Should().Be(3 - 1);
-
-            issue.EndLine.Should().Be(2);
-            issue.EndLineOffset.Should().Be(1 - 1);
-
-            issue.RuleKey.Should().Be("lang1:rule2");
-            issue.FilePath.Should().Be("file");
-            issue.Message.Should().Be("test endline is not zero");
-        }
-
-        [TestMethod]
-        public void ToSonarLintIssue_EndLineIsZero()
-        {
-            // Special case: ignore column offsets if EndLine is zero
-            var ruleConfig = GetDummyRulesConfiguration();
-            var message = new Message("rule3", "ff", 101, 1, 0, 3, "test endline is zero", true, new MessagePart[0]);
-
-            // Act
-            var issue = CFamilyHelper.ToSonarLintIssue(message, "cpp", ruleConfig);
-
-            // Assert
-            issue.StartLine.Should().Be(101);
-
-            issue.EndLine.Should().Be(0);
-            issue.StartLineOffset.Should().Be(0);
-            issue.EndLineOffset.Should().Be(0);
-
-            issue.RuleKey.Should().Be("cpp:rule3");
-            issue.FilePath.Should().Be("ff");
-            issue.Message.Should().Be("test endline is zero");
-        }
-
-        [TestMethod]
-        [DataRow("rule2", AnalysisIssueSeverity.Info, AnalysisIssueType.CodeSmell)]
-        [DataRow("rule3", AnalysisIssueSeverity.Critical, AnalysisIssueType.Vulnerability)]
-        public void ToSonarLintIssue_SeverityAndTypeLookup(string ruleKey, AnalysisIssueSeverity expectedSeverity, AnalysisIssueType expectedType)
-        {
-            var ruleConfig = GetDummyRulesConfiguration();
-
-            var message = new Message(ruleKey, "any", 4, 3, 2, 1, "message", false, new MessagePart[0]);
-            var issue = CFamilyHelper.ToSonarLintIssue(message, "lang1", ruleConfig);
-
-            issue.RuleKey.Should().Be($"lang1:{ruleKey}");
-            issue.Severity.Should().Be(expectedSeverity);
-            issue.Type.Should().Be(expectedType);
-        }
-
-        [TestMethod]
-        [DataRow(IssueSeverity.Blocker, AnalysisIssueSeverity.Blocker)]
-        [DataRow(IssueSeverity.Critical, AnalysisIssueSeverity.Critical)]
-        [DataRow(IssueSeverity.Info, AnalysisIssueSeverity.Info)]
-        [DataRow(IssueSeverity.Major, AnalysisIssueSeverity.Major)]
-        [DataRow(IssueSeverity.Minor, AnalysisIssueSeverity.Minor)]
-        public void ConvertFromIssueSeverity(IssueSeverity cfamilySeverity, AnalysisIssueSeverity analysisIssueSeverity)
-        {
-            CFamilyHelper.Convert(cfamilySeverity).Should().Be(analysisIssueSeverity);
-        }
-
-        [TestMethod]
-        public void ConvertFromIssueSeverity_InvalidValue_Throws()
-        {
-            Action act = () => CFamilyHelper.Convert((IssueSeverity)(-1));
-            act.Should().ThrowExactly<ArgumentOutOfRangeException>().And.ParamName.Should().Be("issueSeverity");
-        }
-
-        [TestMethod]
-        [DataRow(IssueType.Bug, AnalysisIssueType.Bug)]
-        [DataRow(IssueType.CodeSmell, AnalysisIssueType.CodeSmell)]
-        [DataRow(IssueType.Vulnerability, AnalysisIssueType.Vulnerability)]
-        public void ConvertFromIssueType(IssueType cfamilyIssueType, AnalysisIssueType analysisIssueType)
-        {
-            CFamilyHelper.Convert(cfamilyIssueType).Should().Be(analysisIssueType);
-
-            Action act = () => CFamilyHelper.Convert((IssueType) (-1));
-            act.Should().ThrowExactly<ArgumentOutOfRangeException>().And.ParamName.Should().Be("issueType");
-        }
-
-        [TestMethod]
-        public void ConvertFromIssueType_InvalidValue_Throws()
-        {
-            Action act = () => CFamilyHelper.Convert((IssueType)(-1));
-            act.Should().ThrowExactly<ArgumentOutOfRangeException>().And.ParamName.Should().Be("issueType");
-        }
-
         private static ICFamilyRulesConfig GetDummyRulesConfiguration()
         {
             var config = new DummyCFamilyRulesConfig("any")
