+++ src/Integration.Vsix/AsmRef_Integration.Vsix_Baseline_WithoutStrongNames.txt
@@ -1,7 +1,7 @@
 ---
 ################################
 # Assembly references report
-# Report date/time: 2020-08-18T19:50:10.8525256Z
+# Report date/time: 2020-08-19T09:16:17.1507134Z
 ################################
 #
 # Generated by Devtility CheckAsmRefs v0.10.0.217
